% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{icc}
\alias{icc}
\title{Intraclass correlation coefficient (ICC) using mixed effects models}
\usage{
icc(data, ..., group)
}
\arguments{
\item{data}{a data frame containing the data needed for calculated the ICC.}

\item{...}{unquoted variables in the \code{data} argument for which the ICC needs to be calculated.}

\item{group}{an unquoted variable specifying the group by which to calculate the ICC.}
}
\value{
a data frame containing two columns: A \code{Variable} column for which the ICC was calculated and a \code{ICC} which contains the ICC value.
}
\description{
Calculate the intraclass correlation coefficient (ICC) using mixed effects (also known as multilevel) models.
}
