% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{quiet}
\alias{quiet}
\title{Create a quiet function}
\usage{
quiet(.f, silenceMessages = TRUE, silenceWarnings = TRUE)
}
\arguments{
\item{.f}{A function to be silenced}

\item{silenceMessages}{Whether messages should be silenced. Defaults to TRUE.}

\item{silenceWarnings}{Whether warnings should be silenced Defaults to true.}
}
\value{
A wrapped function with  messages and/or warnings suppressed.
If both messages and warnings are suppressed, the function is returned unmodified.
}
\description{
Create a quiet function
}
\examples{
quiet_lmer <- quiet(lme4::lmer)
lme4::lmer(formula = Petal.Width ~ Petal.Length + (Petal.Length | Species), 
           data = iris) # Returns singularity warning
}
