% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{r2}
\alias{r2}
\title{R-squared between or within groups}
\usage{
r2(pred, obs, group = NULL)
}
\arguments{
\item{pred}{a vector containing numeric predictions (i.e. estimated values).}

\item{obs}{a vector containing numeric observations (i.e. truth values).}

\item{group}{An optional group by which to calculate \eqn{R^2}. If left empty. \eqn{R^2} is calculated between groups.}
}
\value{
calculated \eqn{R^2} as a single numeric value.
}
\description{
Function for calculating \eqn{R^2} either within or between groups.
}
\details{
By specifying a some group in the \code{group} argument of this function, \eqn{R^2} is calculated within groups.
Note that this is only meaningful for groups with more than one observation and where variance exists.
}
