% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{se}
\alias{se}
\title{Standard error of predictions between or within groups}
\usage{
se(pred, obs, group = NULL)
}
\arguments{
\item{pred}{a vector containing numeric predictions (i.e. estimated values).}

\item{obs}{a vector containing numeric observations (i.e. truth values).}

\item{group}{An optional group by which to calculate the standard error. If left empty. the standard error is calculated between groups.}
}
\value{
calculated standard error as a single numeric value.
}
\description{
Function for calculating the standard error for predictions either within or between groups.
}
\details{
By specifying a some group in the \code{group} argument of this function, the standard error is calculated within groups.
Note that this is only meaningful for groups with more than one observation and where variance exists.
}
